requirements.txt
Django==1.10
argparse==1.2.1
django-filter==0.14.0
djangorestframework==3.4.6
wsgiref==0.1.2
